re-structure test directory a bit
