+++ hoyo.cabal
@@ -79,6 +79,9 @@ executable hoyo-cli
 
 test-suite hoyo-test
     main-is:          Main.hs
+    other-modules:    HoYo.Test.Gen
+                      HoYo.Test.Bookmark
+    -- other-extensions:
     hs-source-dirs:   test
     default-language: Haskell2010
     ghc-options:      -Wall
@@ -90,7 +93,5 @@ test-suite hoyo-test
         filepath >=1.4 && <1.5,
         time >=1.12 && <1.13,
         hoyo
-    other-modules:    Gen
-    -- other-extensions:
     type:             exitcode-stdio-1.0
     default-extensions: OverloadedStrings
